P2P Gameloop networking
Sending messages between gamesteps in peer-to-peer lockstep

+++ trenk/Assets/Scripts/Online/INetSerializer.cs
@@ -7,5 +7,5 @@ public interface INetSerializer
     /// <summary>
     /// Handles received data over network
     /// </summary>
-    bool Receive(short type, byte[] data);
+    bool Receive(byte type, byte[] data);
 }
